+++ Resources/Locale/ru-RU/Imperial/Medieval/medieval.ftl
@@ -4465,7 +4465,7 @@ ent-MedievalTelega = телега с припасами
 ent-MedievalCart = телега
     .desc = Позволяет перевозить груз, почти не замедляет. Грабить корованы
     .suffix = { "Средневековье" }
-ent-MedievalCartTemplate = заготовка телеги
+ent-MedievalCartCraft = заготовка телеги
     .desc = Позволяет перевозить груз, почти не замедляет.
     .suffix = { "Средневековье" }
 ent-MedievalKaravanSkeletonSpawner = спавнер проклятого защитника корована
