+++ Resources/Prototypes/Imperial/Medieval/ratling.yml
@@ -409,6 +409,13 @@
 - type: inventoryTemplate
   id: ratling
   slots:
+    - name: shoes
+      slotTexture: shoes
+      slotFlags: FEET
+      stripTime: 3
+      uiWindowPos: 1,0
+      strippingWindowPos: 1,3
+      displayName: Shoes
     - name: jumpsuit
       slotTexture: uniform
       slotFlags: INNERCLOTHING
